Moving the ADM struct out of Get_Lambda
Makes more sense in its own file
